+++ packages/backend/src/projectAdapters/dbtBaseProjectAdapter.ts
@@ -178,6 +178,7 @@ export class DbtBaseProjectAdapter implements ProjectAdapter {
                     loadSources,
                     adapterType,
                     metrics,
+                    this.warehouseClient.getStartOfWeek(),
                 );
                 return [...explores, ...failedExplores];
             }