+++ packages/backend/src/services/services.ts
@@ -73,11 +73,13 @@ export const healthService = new HealthService({
 
 export const dashboardService = new DashboardService({
     dashboardModel,
+    spaceModel,
 });
 
 export const savedChartsService = new SavedChartService({
     projectModel,
     savedChartModel,
+    spaceModel,
 });
 
 export const personalAccessTokenService = new PersonalAccessTokenService({