+++ packages/frontend/src/components/common/Filters/FiltersProvider.tsx
@@ -1,4 +1,4 @@
-import { FilterableField, FilterRule } from '@lightdash/common';
+import { FilterableField, FilterRule, WeekDay } from '@lightdash/common';
 import React, { createContext, FC, useCallback, useContext } from 'react';
 
 export type FieldWithSuggestions = FilterableField & {
@@ -10,6 +10,7 @@ export type FieldsWithSuggestions = Record<string, FieldWithSuggestions>;
 type FiltersContext = {
     projectUuid: string;
     fieldsMap: FieldsWithSuggestions;
+    startOfWeek?: WeekDay | null;
     getField: (filterRule: FilterRule) => FieldWithSuggestions | undefined;
 };
 
@@ -18,11 +19,13 @@ const Context = createContext<FiltersContext | undefined>(undefined);
 type Props = {
     projectUuid: string;
     fieldsMap: Record<string, FieldWithSuggestions>;
+    startOfWeek?: WeekDay | null;
 };
 
 export const FiltersProvider: FC<Props> = ({
     projectUuid,
     fieldsMap,
+    startOfWeek,
     children,
 }) => {
     const getField = useCallback(
@@ -34,7 +37,9 @@ export const FiltersProvider: FC<Props> = ({
         [fieldsMap],
     );
     return (
-        <Context.Provider value={{ projectUuid, fieldsMap, getField }}>
+        <Context.Provider
+            value={{ projectUuid, fieldsMap, startOfWeek, getField }}
+        >
             {children}
         </Context.Provider>
     );