+++ packages/frontend/src/pages/Dashboard.tsx
@@ -1,4 +1,4 @@
-import { Alert, Intent, Spinner } from '@blueprintjs/core';
+import { Alert, Intent, NonIdealState, Spinner } from '@blueprintjs/core';
 import { Dashboard as IDashboard, DashboardTileTypes } from '@lightdash/common';
 import React, {
     FC,
@@ -9,7 +9,7 @@ import React, {
     useState,
 } from 'react';
 import { Layout, Responsive, WidthProvider } from 'react-grid-layout';
-import { useHistory, useParams } from 'react-router-dom';
+import { Redirect, useHistory, useParams } from 'react-router-dom';
 import DashboardHeader from '../components/common/Dashboard/DashboardHeader';
 import Page from '../components/common/Page/Page';
 import DashboardFilter from '../components/DashboardFilter';
@@ -49,12 +49,25 @@ const GridTile: FC<
         tile.type === DashboardTileTypes.SAVED_CHART
             ? tile.properties?.savedChartUuid || undefined
             : undefined;
-    const { data: savedQuery, isLoading } = useSavedQuery({
+    const {
+        data: savedQuery,
+        isLoading,
+        isError,
+    } = useSavedQuery({
         id: savedChartUuid,
     });
     switch (tile.type) {
         case DashboardTileTypes.SAVED_CHART:
             if (isLoading) return <></>;
+            if (isError)
+                return (
+                    <TileBase title={''} {...props}>
+                        <NonIdealState
+                            icon="lock"
+                            title={`You don't have access to view this chart`}
+                        ></NonIdealState>
+                    </TileBase>
+                );
             return (
                 <UnderlyingDataProvider
                     filters={savedQuery?.metricQuery.filters}
@@ -103,7 +116,8 @@ const Dashboard = () => {
         [dashboardTemporaryFilters],
     );
     const isEditMode = useMemo(() => mode === 'edit', [mode]);
-    const { data: dashboard } = useDashboardQuery(dashboardUuid);
+    const { data: dashboard, error: dashboardError } =
+        useDashboardQuery(dashboardUuid);
     const [hasTilesChanged, setHasTilesChanged] = useState<boolean>(false);
     const [dashboardName, setDashboardName] = useState<string>('');
     const {
@@ -306,10 +320,12 @@ const Dashboard = () => {
         dashboardUuid,
     ]);
 
+    if (dashboardError) {
+        return <Redirect to="/no-access" />;
+    }
     if (dashboard === undefined) {
         return <Spinner />;
     }
-
     const dashboardChartTiles = dashboardTiles.filter(
         (tile) => tile.type === DashboardTileTypes.SAVED_CHART,
     );